+++ msg/src/pages/Club/AllPosts.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components"
-import Header from "../../components/header"
-import Footer from "../../components/footer"
+import Header from "../../components/common/header"
+import Footer from "../../components/common/footer"
 import Post from "../../components/club/Post"
 import PostContainer from "../../components/club/PostContainer"
 import Before from "../../components/Before"
