+++ msg/src/pages/Club/ClubPage.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components"
 import { Colors } from "../../styles/colors"
-import Header from "../../components/header"
-import Footer from "../../components/footer"
+import Header from "../../components/common/header"
+import Footer from "../../components/common/footer"
 import { FaBell } from "react-icons/fa6"
 import { IoIosArrowForward } from "react-icons/io"
 import Post from "../../components/club/Post"
