+++ msg/src/pages/Inform/InformPage.tsx
